Move UI to separate file
